refactor(components): bar in separate directory
